rename tracker's model folder to trackers
