+++ perception/autoware_multi_object_tracker/include/autoware/multi_object_tracker/tracker/tracker.hpp
@@ -15,11 +15,11 @@
 #ifndef AUTOWARE__MULTI_OBJECT_TRACKER__TRACKER__TRACKER_HPP_
 #define AUTOWARE__MULTI_OBJECT_TRACKER__TRACKER__TRACKER_HPP_
 
-#include "model/multiple_vehicle_tracker.hpp"
-#include "model/pedestrian_and_bicycle_tracker.hpp"
-#include "model/pedestrian_tracker.hpp"
-#include "model/polygon_tracker.hpp"
-#include "model/tracker_base.hpp"
-#include "model/vehicle_tracker.hpp"
+#include "trackers/multiple_vehicle_tracker.hpp"
+#include "trackers/pedestrian_and_bicycle_tracker.hpp"
+#include "trackers/pedestrian_tracker.hpp"
+#include "trackers/polygon_tracker.hpp"
+#include "trackers/tracker_base.hpp"
+#include "trackers/vehicle_tracker.hpp"
 
 #endif  // AUTOWARE__MULTI_OBJECT_TRACKER__TRACKER__TRACKER_HPP_
