+++ perception/autoware_multi_object_tracker/lib/tracker/trackers/multiple_vehicle_tracker.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "autoware/multi_object_tracker/tracker/model/multiple_vehicle_tracker.hpp"
+#include "autoware/multi_object_tracker/tracker/trackers/multiple_vehicle_tracker.hpp"
 
 #include "autoware/multi_object_tracker/object_model/object_model.hpp"
